abstracing nav item to its own component
