+++ src/components/Organisms/Header.tsx
@@ -1,6 +1,7 @@
 import { withPrefix } from 'gatsby'
 import React, { FC } from 'react'
 import styled from 'styled-components'
+import NavItem from '../../components/Molecules/NavItem'
 import Container from '../Atoms/Container'
 import Link from '../Atoms/Link'
 import Wrapper from '../Atoms/Wrapper'
@@ -66,15 +67,7 @@ const Header: FC<IHeaderProps> = ({ primaryNav, siteTitle }) => (
     <HeaderContainer>
       <Link to={withPrefix('/')}>{siteTitle}</Link>
       <MainNav>
-        {primaryNav
-          ? primaryNav.edges.map(item => (
-              <li key={item.node.frontmatter.path}>
-                <Link activeClassName="active" to={withPrefix(item.node.frontmatter.path)}>
-                  {item.node.frontmatter.title}
-                </Link>
-              </li>
-            ))
-          : null}
+        {primaryNav ? primaryNav.edges.map(item => NavItem(item)) : null}
         <li>
           <Link to="/post" activeClassName="active">
             Posts
